Add more tests
As part of porting ArnoldC to javascript, I wrote some additional tests
to help determine what the correct behaviour is in some cases.  These
tests are mostly focused around how variable scope works, numeric
limits, and other edge cases.

The only existing test I modified was in the logical test, and that was
just to change its name (the name said it was expecting a result of
False but it really expected a result of True).